+++ core/processor/processor.go
@@ -201,6 +201,9 @@ type Banking interface {
 	BridgeStopped(context.Context, bool, string, uint64, uint64, string) error
 	BridgeResumed(context.Context, bool, string, uint64, uint64, string) error
 	CheckTransfer(t *types.TransferBase) error
+	NewGovernanceTransfer(ctx context.Context, ID, reference string, transferConfig *types.NewTransferConfiguration) error
+	VerifyGovernanceTransfer(transfer *types.NewTransferConfiguration) error
+	VerifyCancelGovernanceTransfer(transferID string) error
 }
 
 // NetworkParameters ...